Hypergeometric Log Likelihood Function

The log likelihood of a hypergeometric density with data, x, m, n and k parameters.

likhyper(x, m, n, k, log = TRUE)

Arguments

x
vector of quantiles representing the number of white balls drawn without replacement from an urn which contains both black and white balls.

m
the number of white balls in the urn.

n
the number of black balls in the urn.

k
the number of balls drawn from the urn.

log
logical; if TRUE, probabilities p are given as log(p).

Value

A numeric scalar for the log likelihood of the hypergeometric density given the data where m, n, and k can be held constant or if vector were given vector will be returned.

Details

The log likelihood is the log of a function of parameters given the data.

Examples

likhyper(x = rhyper(nn = 2, m = 3, n = 4, k = 2),
        m = 3, n = 4, k = 2)

#> [1] -1.119232